AAE-4241 Fix unit tests
