+++ lib/process-services-cloud/src/lib/form/components/widgets/date/date-cloud.widget.spec.ts
@@ -21,7 +21,6 @@ import { setupTestBed, FormFieldModel, FormModel } from '@alfresco/adf-core';
 import moment from 'moment-es6';
 import { ProcessServiceCloudTestingModule } from '../../../../testing/process-service-cloud.testing.module';
 import { TranslateModule } from '@ngx-translate/core';
-import { SimpleChanges } from '@angular/core';
 
 describe('DateWidgetComponent', () => {
 
@@ -221,17 +220,8 @@ describe('DateWidgetComponent', () => {
                 const dateElement: any = element.querySelector('#date-field-id');
                 expect(dateElement.value).toContain('12-30-9999');
 
-                const newField = { ...field, value: '03-02-2020' };
+                widget.field.value = '03-02-2020';
 
-                const changes: SimpleChanges = {
-                    'field': {
-                        previousValue: field,
-                        currentValue: newField,
-                        firstChange: false,
-                        isFirstChange(): boolean { return this.firstChange; }
-                    }
-                };
-                widget.ngOnChanges(changes);
                 fixture.detectChanges();
                 fixture.whenStable()
                     .then(() => {
@@ -240,70 +230,4 @@ describe('DateWidgetComponent', () => {
             });
     });
 
-    it('should not call on change when is first change or field is not set or the field value does not change', () => {
-        const field = new FormFieldModel(new FormModel(), {
-            id: 'date-field-id',
-            name: 'date-name',
-            value: '12-30-9999',
-            type: 'date',
-            readOnly: 'false'
-        });
-        field.isVisible = true;
-        field.dateDisplayFormat = 'MM-DD-YYYY';
-        widget.field = field;
-        widget.ngOnInit();
-        fixture.detectChanges();
-        fixture.whenStable()
-            .then(() => {
-                expect(element.querySelector('#date-field-id')).toBeDefined();
-                expect(element.querySelector('#date-field-id')).not.toBeNull();
-                const dateElement: any = element.querySelector('#date-field-id');
-                expect(dateElement.value).toContain('12-30-9999');
-
-                const newField = { ...field, value: '03-02-2020' };
-
-                let changes: SimpleChanges = {
-                    'field': {
-                        previousValue: field,
-                        currentValue: newField,
-                        firstChange: true,
-                        isFirstChange(): boolean { return this.firstChange; }
-                    }
-                };
-                widget.ngOnChanges(changes);
-                fixture.detectChanges();
-                fixture.whenStable()
-                    .then(() => {
-                        expect(dateElement.value).toContain('12-30-9999');
-                        changes = {};
-                        widget.ngOnChanges(changes);
-                        fixture.detectChanges();
-                        fixture.whenStable()
-                            .then(() => {
-                                expect(dateElement.value).toContain('12-30-9999');
-                                changes = {
-                                    'field': {
-                                        previousValue: field,
-                                        currentValue: field,
-                                        firstChange: false,
-                                        isFirstChange(): boolean { return this.firstChange; }
-                                    }
-                                };
-                                widget.ngOnChanges(changes);
-                                fixture.detectChanges();
-                                fixture.whenStable()
-                                    .then(() => {
-                                        expect(dateElement.value).toContain('12-30-9999');
-                                        changes = null;
-                                        widget.ngOnChanges(changes);
-                                        fixture.detectChanges();
-                                        fixture.whenStable()
-                                            .then(() => {
-                                                expect(dateElement.value).toContain('12-30-9999');
-                                            });
-                                    });
-                            });
-                    });
-            });
-    });
 });
